task/TRAIN-1084 Refactor: Refactor Home Page and other Components separated
