+++ src/components/Buttons/AddButton.stories.tsx
@@ -1,0 +1,19 @@
+import { StoryObj, Meta } from '@storybook/react';
+import { action } from '@storybook/addon-actions';
+import AddButton, { AddButtonProps } from './AddButton';
+
+const meta: Meta<AddButtonProps> = {
+  title: 'Components/AddButton',
+  component: AddButton,
+  argTypes: {
+    onClick: { action: 'clicked' },
+  },
+};
+
+export default meta;
+type Story = StoryObj<AddButtonProps>;
+
+export const Default: Story = {
+  args: {},
+};
+
