+++ src/features/home/components/elements.tsx
@@ -1,0 +1,6 @@
+import styled from 'styled-components';
+
+export const AppContainer = styled.div`
+  flex-direction: column;
+  margin-top: 20px;
+`;
