+++ src/features/home/routes/HomePage.tsx
@@ -1,0 +1,11 @@
+import Home from '../components/Home';
+
+const HomePage = () => {
+  return (
+    <div>
+      <Home />
+    </div>
+  );
+};
+
+export default HomePage;
