+++ src/features/home/routes/index.tsx
@@ -1,0 +1,11 @@
+import { RouteObject } from 'react-router-dom';
+import Home from '../components/Home';
+
+const HomeRoutes: RouteObject[] = [
+  {
+    path: '/',
+    element: <Home />,
+  },
+];
+
+export default HomeRoutes;
